# swagger.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: BackToME API
  description: API para busca de animais perdidos e adoção de animais.

servers:
  - url: http://localhost:3000
    description: Servidor local

  - url: https://api-backtome.onrender.com
    description: Servidor Render

tags:
- name: Validações
- name: Usuários
- name: Animais Perdidos
- name: Animais para Adoção
- name: ONGs

paths:
  /validate/email:
    post:
      tags:
        - Validações
      summary: Validar email
      description: Verifica se o usuário existe no banco de dados
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidateEmail"
            examples:
              Example Json:
                value:
                  email: user@email.com
              Example Type:
                value:
                  email: string
      responses:
        "200":
          description: Não existe nenhum usuário cadastrado com este email
        "400":
          description: Já existe um cadastro com esse email

  /users/registered:
    post:
      tags:
      - Usuários
      summary: Cadastrar novo usuário
      description: Rota para cadastrar um novo usuário.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegistered"
            examples:
              Exemple Json:
                value:
                  name: "Nome do usuário"
                  email: "email@example.com"
                  cpf: "12345678900"
                  phone: "123456789"
                  password: "senha123"
                  address:
                    cep: "25264253"
                    complement: "502"

              Exemple Type:
                value:
                  name: "string"
                  email: "string"
                  cpf: "string"
                  phone: "string"
                  password: "string"
                  address:
                    cep: "string"
                    complement: "string"
      responses:
        "200":
          description: Usuário cadastrado com sucesso!
        "400":
          description: Erro ao cadastrar usuário!
  /users/login:
    post:
      tags:
      - Usuários
      summary: Login usuário
      description: Rota login do usuário.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserLogin"
            examples:
              Example Json:
                value:
                  email: "email@example.com"
                  password: "123456"
              Example Type:
                value:
                  email: "string"
                  password: "string"

      responses:
        "200":
          description: Usuário cadastrado com sucesso!
        "400":
          description: Erro ao cadastrar usuário!
  /users/{userId}:
    put:
      tags:
      - Usuários
      summary: Editar dados do usuário
      description: Rota para editar dados do usuário.
      parameters:
      - name: userId
        in: path
        required: true
        description: ID do usuário
        schema:
          type: integer
          format: int64
          example: 123
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserEdited"
            examples:
              Example Json:
                value:
                  name: "Nome do usuário"
                  email: "email@example.com"
                  cpf: "12345678900"
                  phone: "123456789"
                  password: "senha123"
                  address:
                    cep: "25264253"
                    complement: "502"
              Exemple Type:
                value:
                  name: "string"
                  email: "string"
                  cpf: "string"
                  phone: "string"
                  password: "string"
                  address:
                    cep: "string"
                    complement: "string"
      responses:
        "200":
          description: Usuário cadastrado com sucesso!
        "400":
          description: Erro interno!

  /users/{userId}/find:
    get:
      tags:
      - Usuários
      summary: Buscar dados do usuário
      description: Rota para listar dados do usuário.
      parameters:
      - name: userId
        in: path
        required: true
        description: ID do usuário
        schema:
          type: integer
          format: int64
          example: 123
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserList"
            examples:
              Example Json:
                value:
                  name: "Nome do usuário"
                  email: "email@example.com"
                  cpf: "12345678900"
                  phone: "123456789"
                  password: "senha123"
                  address:
                    cep: "25264253"
                    complement: "502"
              Exemple Type:
                value:
                  name: "string"
                  email: "string"
                  cpf: "string"
                  phone: "string"
                  password: "string"
                  address:
                    cep: "string"
                    complement: "string"
      responses:
        "200":
          description: Usuário cadastrado com sucesso!
        "400":
          description: Erro interno!

  /users/recovery-password:
    post:
      tags:
        - Usuários
      summary: Recuperar senha
      description: Rota para recuperar senha
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecoveryPassword"
            examples:
              Example Json:
                value:
                  email: user@email.com
                  cpf: 111.111.111-11
              Example Type:
                value:
                  email: string
                  cpf: string
      responses:
        "200":
          description: Link para verificação

        "400":
          description: Erro recuperar senha
  /users/change-password/{userId}:
    post:
      tags:
        - Usuários
      summary: Alterar a senha
      description: Rota para recuperar senha
      parameters:
      - name: userId
        in: path
        required: true
        description: ID do usuário
        schema:
          type: integer
          format: int64
          example: 123
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePassword"
            examples:
              Example Json:
                value:
                  password: A123456@
                  newPassword: B789456,
              Example Type:
                value:
                  password: string
                  newPassword: string
      responses:
        "200":
          description: Senha alterada com sucesso

        "400":
          description: Erro ao alterar senha

  /animals-found/registered:
    post:
      tags:
      - Animais Encontrados
      summary: Cadastrar novo animal encontrado
      description: Rota para cadastrar um novo animal encontrado.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnimalsFound"
            examples:
              Example Json:
                value:
                  photo: "Caminho para a foto do animal"
                  local_encontrado: "Descrição ou endereço do local encontrado"
                  species: "Espécie do animal"
                  race: "Raça do animal"
                  age: "Filhote || adulto || idoso"
                  color: "Cor do animal"
                  size: "Tamanho do animal"
                  distinctive_characteristics: "Características distintivas do animal"
                  user_id: "usuario id"
              Example Type:
                value:
                  photo: "string"
                  local_encontrado: "string"
                  species: "string"
                  race: "string"
                  age: "string"
                  color: "string"
                  size: "string"
                  distinctive_characteristics: "string"
                  user_id: "string"
      responses:
        "200":
          description: Animal cadastrado com sucesso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnimalsFound"
              examples:
                Example Json:
                  value:
                    id: 1
                    foto: "Caminho para a foto do animal"
                    local_encontrado: "Descrição ou endereço do local encontrado"
                    species: "Espécie do animal"
                    race: "Raça do animal"
                    age: "Filhote || adulto || idoso"
                    color: "Cor do animal"
                    size: "Tamanho do animal"
                    distinctive_characteristics: "Características distintivas do animal"
                    user_id: "usuario id"
                Example Type:
                  value:
                    photo: "string"
                    local_encontrado: "string"
                    species: "string"
                    race: "string"
                    age: "string"
                    color: "string"
                    size: "string"
                    distinctive_characteristics: "string"
                    user_id: "string"
        "400":
          description: Erro ao cadastrar animal
  /lost-animals/registered:
    post:
      tags:
      - Animais Perdidos
      summary: Cadastrar novo animal perdido
      description: Rota para cadastrar um novo animal perdido.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LostAnimals"
            examples:
              Example Json:
                value:
                  species: "Espécie do animal"
                  race: "Raça do animal"
                  age: "Filhote || adulto || idoso"
                  color: "Cor do animal"
                  size: "Tamanho do animal"
                  distinctive_characteristics: "Características distintivas do animal"
                  date_loss: "Data da perda"
                  location_loss: "Local onde o animal foi perdido"
                  user_id: "usuario id"
              Example Type:
                value:
                  photo: "string"
                  local_encontrado: "string"
                  species: "string"
                  race: "string"
                  age: "string"
                  color: "string"
                  size: "string"
                  distinctive_characteristics: "string"
                  user_id: "string"

      responses:
        "200":
          description: Animal perdido cadastrado com sucesso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LostAnimals"
              examples:
                Example Json:
                  value:
                    id: 1
                    species: "Espécie do animal"
                    race: "Raça do animal"
                    age: "Filhote || adulto || idoso"
                    color: "Cor do animal"
                    size: "Tamanho do animal"
                    distinctive_characteristics: "Características distintivas do animal"
                    date_loss: "Data da perda"
                    location_loss: "Local onde o animal foi perdido"
                    user_id: "usuario id"
                Example Type:
                  value:
                    photo: "string"
                    local_encontrado: "string"
                    species: "string"
                    race: "string"
                    age: "string"
                    color: "string"
                    size: "string"
                    distinctive_characteristics: "string"
                    user_id: "string"
        "400":
          description: Erro ao cadastrar animal perdido
  /animals-adoption/registered:
    post:
      tags:
      - Animais para Adoção
      summary: Cadastrar novo animal para adoção
      description: Rota para cadastrar um novo animal disponível para adoção.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AvailableAnimals"
            examples:
              Example Json:
                value:
                  species: "Espécie do animal"
                  race: "Raça do animal"
                  age: "Filhote || adulto || idoso"
                  color: "Cor do animal"
                  size: "Tamanho do animal"
                  distinctive_characteristics: "Características distintivas do animal"
                  informacoes_personalidade: "Informações sobre a personalidade do animal"
                  user_id: "id usuario"
              Example Type:
                value:
                  photo: "string"
                  local_encontrado: "string"
                  species: "string"
                  race: "string"
                  age: "string"
                  color: "string"
                  size: "string"
                  distinctive_characteristics: "string"
                  user_id: "string"
      responses:
        "200":
          description: Animal para adoção cadastrado com sucesso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailableAnimals"
              examples:
                Example Json:
                  value:
                    id: 1
                    species: "Espécie do animal"
                    race: "Raça do animal"
                    age: "Filhote"
                    color: "Cor do animal"
                    size: "Tamanho do animal"
                    distinctive_characteristics: "Características distintivas do animal"
                    informacoes_personalidade: "Informações sobre a personalidade do animal"
                    user_id: "id usuario"
                Example Type:
                  value:
                    photo: "string"
                    local_encontrado: "string"
                    species: "string"
                    race: "string"
                    age: "string"
                    color: "string"
                    size: "string"
                    distinctive_characteristics: "string"
                    user_id: "string"
        "400":
          description: Erro ao cadastrar animal para adoção
  /available-animal/find:
    get:
      tags:
      - Animais para Adoção
      summary: Buscar animais para adoção
      description: Rota para buscar animais disponíveis para adoção.
      parameters:
      - in: query
        name: especie
        schema:
          type: string
        description: Espécie do animal
      - in: query
        name: raca
        schema:
          type: string
        description: Raça do animal
      - in: query
        name: cor
        schema:
          type: string
        description: Cor do animal
      - in: query
        name: tamanho
        schema:
          type: string
        description: Tamanho do animal
      - in: query
        name: caracteristicas_distintivas
        schema:
          type: string
        description: Características distintivas do animal
      responses:
        "200":
          description: Lista de animais para adoção encontrada
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AvailableAnimalsFind"
              examples:
                Example Json:
                  value:
                  - id: 1
                    species: "Espécie do animal"
                    race: "Raça do animal"
                    age: "Filhote"
                    color: "Cor do animal"
                    size: "Tamanho do animal"
                    distinctive_characteristics: "Características distintivas do animal"
                    informacoes_personalidade: "Informações sobre a personalidade do animal"
                    user_id: "id usuario"
                  - id: 2
                    species: "Espécie do animal"
                    race: "Raça do animal"
                    age: "Filhote || adulto || idoso"
                    color: "Cor do animal"
                    size: "Tamanho do animal"
                    distinctive_characteristics: "Características distintivas do animal"
                    informacoes_personalidade: "Informações sobre a personalidade do animal"
                    user_id: "id usuario"
        "400":
          description: Erro ao buscar animais para adoção
  /lost-animals/find:
    get:
      tags:
      - Animais Perdidos
      summary: Buscar animais perdidos
      description: Rota para buscar animais perdidos
      parameters:
      - in: query
        name: especie
        schema:
          type: string
        description: Espécie do animal
      - in: query
        name: raca
        schema:
          type: string
        description: Raça do animal
      - in: query
        name: cor
        schema:
          type: string
        description: Cor do animal
      - in: query
        name: tamanho
        schema:
          type: string
        description: Tamanho do animal
      - in: query
        name: caracteristicas_distintivas
        schema:
          type: string
        description: Características distintivas do animal
      responses:
        "200":
          description: Lista de animais para adoção encontrada
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LostAnimalsFind"
              examples:
                Example Json:
                  value:
                  - id: 1
                    species: "Espécie do animal"
                    race: "Raça do animal"
                    age: "Filhote"
                    color: "Cor do animal"
                    size: "Tamanho do animal"
                    distinctive_characteristics: "Características distintivas do animal"
                    informacoes_personalidade: "Informações sobre a personalidade do animal"
                    user_id: "id usuario"
                  - id: 2
                    species: "Espécie do animal"
                    race: "Raça do animal"
                    age: "Filhote || adulto || idoso"
                    color: "Cor do animal"
                    size: "Tamanho do animal"
                    distinctive_characteristics: "Características distintivas do animal"
                    informacoes_personalidade: "Informações sobre a personalidade do animal"
                    user_id: "id usuario"
        "400":
          description: Erro ao buscar animais perdido
  /ongs:
    get:
      tags:
        - ONGs
      summary: Listar todas as ONGs
      description: Rota para listar todas as Organizações Não Governamentais (ONGs).
      responses:
        "200":
          description: Lista de ONGs encontrada
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ongs"
        "404":
          description: ONG não encontrada

components:
  schemas:
    UserRegistered:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        cpf:
          type: string
        phone:
          type: string
        password:
          type: string
        address:
          type: object
          properties:
            cep:
              type: string
            complement:
              type: string

    UserLogin:
      type: object
      properties:
        email:
          type: string
        password:
          type: string

    UserEdited:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        cpf:
          type: string
        phone:
          type: string
        password:
          type: string
        address:
          type: object
          properties:
            cep:
              type: string
            complement:
              type: string

    UserList:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        cpf:
          type: string
        phone:
          type: string
        password:
          type: string
        address:
          type: object
          properties:
            cep:
              type: string
            complement:
              type: string

    RecoveryPassword:
      type: object
      properties:
        email:
          type: string
        password:
          type: string

    ChangePassword:
      type: object
      properties:
        email:
          type: string
        password:
          type: string

    ValidateEmail:
      type: object
      properties:
        email:
          type: string

    AnimalsFound:
      type: object
      properties:
        photo:
          type: string
        local_encontrado:
          type: string
        species:
          type: string
        race:
          type: string
        age:
          type: string
        color:
          type: string
        size:
          type: string
        distinctive_characteristics:
          type: string
        user_id:
          type: string

    LostAnimals:
      type: object
      properties:
        species:
          type: string
        race:
          type: string
        age:
          type: string
        color:
          type: string
        size:
          type: string
        distinctive_characteristics:
          type: string
        date_loss:
          type: string
        location_loss:
          type: string
        user_id:
          type: string

    AvailableAnimals:
      type: object
      properties:
        species:
          type: string
        race:
          type: string
        age:
          type: string
        color:
          type: string
        size:
          type: string
        distinctive_characteristics:
          type: string
        informacoes_personalidade:
          type: string
        user_id:
          type: string

    LostAnimalsFind:
      type: object
      properties:
        id:
          type: integer
        species:
          type: string
        race:
          type: string
        age:
          type: string
        color:
          type: string
        size:
          type: string
        distinctive_characteristics:
          type: string
        informacoes_personalidade:
          type: string
        user_id:
          type: string

    AvailableAnimalsFind:
      type: object
      properties:
        id:
          type: integer
        species:
          type: string
        race:
          type: string
        age:
          type: string
        color:
          type: string
        size:
          type: string
        distinctive_characteristics:
          type: string
        informacoes_personalidade:
          type: string
        user_id:
          type: string

    Ongs:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        address:
          type: string
        contact:
          type: string
        website:
          type: string
        description:
          type: string
        logo:
          type: string